fix(#S11P12D203-198): 게임01 관련 ID값을 모두 Long 타입으로 변경

+++ android/project/app/src/main/java/com/chill/mallang/data/api/QuizApi.kt
@@ -14,12 +14,12 @@ import retrofit2.http.Path
 interface QuizApi {
     @GET("quiz/start/{areaId}")
     suspend fun getQuizIds(
-        @Path("areaId") areaId: Int,
-    ): Response<ResponseBody<List<Int>>>
+        @Path("areaId") areaId: Long,
+    ): Response<ResponseBody<List<Long>>>
 
     @GET("quiz/{quizId}")
     suspend fun getQuiz(
-        @Path("quizId") quizId: Int,
+        @Path("quizId") quizId: Long,
     ): Response<ResponseBody<Game01QuizData>>
 
     @POST("quiz/submit")
